# avoid ever copying plotpoints