#### PR Type

- [ ] 💰 Feature
- [ ] 🪲 Bug Fix
- [x] 👷‍♀️ Refactor
- [ ] 📖 Documentation Update

#### Description

<!--- Provide a summary of your changes -->
<!--- Link to any related issues this PR resolves -->
<!--- If integration tests have been performed, describe the tests along with their outcomes -->

Resolves https://github.com/luftkode/plotinator3000/issues/190

#### Checklist

- [x] 🌞 Changelog updated