@CramBL 2025-02-14
# Profiling

## Before

### Downsampling (~60 ms)

![{AEB22A2E-3C0D-4C8A-85BA-C97A17A18679}](https://github.com/user-attachments/assets/c16de27f-13ee-459e-887e-ddc45c79498f)

### Raw (~520 ms)

![{DA7836EF-3FB6-46D8-88F9-F7F26D68A9FC}](https://github.com/user-attachments/assets/f57b7d87-7615-4aeb-8f46-da26701dfc35)

### Downsampling /w zoom (~29 ms)

![{B3854BCD-410D-4CC1-8DE8-B60F62700DCF}](https://github.com/user-attachments/assets/71275718-7346-4c67-aca7-e5b863b6a8ea)

### Raw /w zoom (~59 ms)

![{5B8ED90D-36AC-45E2-901C-28A7ADDA464B}](https://github.com/user-attachments/assets/23905f81-c0bf-4b4f-a214-2a92322a284c)

## After

### Downsampling (~50 ms)

![{CEBCBD2A-3C5A-4E45-B1F9-F41514D305D7}](https://github.com/user-attachments/assets/840bc0da-5bbf-4264-bbe0-49744fe75389)


### Raw (~500ms)

![{B3ED88F1-DACE-49F4-A140-DC5EB358790A}](https://github.com/user-attachments/assets/9fbc4e3a-75c1-453a-8f78-1dd1d3bc9eb7)

### Downsampling w/ zoom (~27 ms)

![{3053DA75-C345-4A09-806C-00732DD414CA}](https://github.com/user-attachments/assets/a81e8ae7-9867-4e53-a499-185898310c83)


### Raw w/ zoom (~65 ms)

![{9D949B4B-7A18-4341-A080-6886710219FB}](https://github.com/user-attachments/assets/00dcad4e-58e5-49d1-b0f8-1e081b72f335)

## Summary

Code execution before the tesselation step was already very fast, now it's like ~20% faster. No more needless copying when slicing into plotpoints 👍 .